Add Upcoming & NewestViewModel
