+++ app/src/main/java/com/yesferal/hornsapp/app/presentation/di/PresentationModule.kt
@@ -2,6 +2,8 @@ package com.yesferal.hornsapp.app.presentation.di
 
 import com.yesferal.hornsapp.app.presentation.ui.band.BandViewModelFactory
 import com.yesferal.hornsapp.app.presentation.ui.concert.detail.ConcertViewModelFactory
+import com.yesferal.hornsapp.app.presentation.ui.concert.newest.NewestViewModelFactory
+import com.yesferal.hornsapp.app.presentation.ui.concert.upcoming.UpcomingViewModelFactory
 import com.yesferal.hornsapp.app.presentation.ui.favorite.FavoritesViewModelFactory
 import com.yesferal.hornsapp.app.presentation.ui.home.HomeViewModelFactory
 import com.yesferal.hornsapp.app.presentation.ui.main.MainViewModelFactory
@@ -27,6 +29,18 @@ fun Container.registerPresentationModule() {
     this register Factory {
         FavoritesViewModelFactory(
             getFavoriteConcertsUseCase = resolve()
+        )
+    }
+
+    this register Factory {
+        UpcomingViewModelFactory(
+            getConcertsUseCase = resolve()
+        )
+    }
+
+    this register Factory {
+        NewestViewModelFactory(
+            getConcertsUseCase = resolve()
         )
     }
