+++ app/src/main/java/com/yesferal/hornsapp/app/presentation/ui/concert/newest/NewestFragment.kt
@@ -11,8 +11,6 @@ import com.yesferal.hornsapp.app.presentation.common.base.ParcelableViewData
 import com.yesferal.hornsapp.app.presentation.common.custom.*
 import com.yesferal.hornsapp.app.presentation.ui.concert.upcoming.UpcomingViewData
 import com.yesferal.hornsapp.app.presentation.ui.home.HomeFragmentDirections
-import com.yesferal.hornsapp.app.presentation.ui.home.HomeViewModel
-import com.yesferal.hornsapp.app.presentation.ui.home.HomeViewModelFactory
 import com.yesferal.hornsapp.multitype.MultiTypeAdapter
 import com.yesferal.hornsapp.multitype.model.ViewHolderBinding
 import kotlinx.android.synthetic.main.fragment_newest.*
@@ -24,7 +22,7 @@ class NewestFragment
         get() = R.layout.fragment_newest
 
     private lateinit var multiTypeAdapter: MultiTypeAdapter
-    private lateinit var homeViewModel: HomeViewModel
+    private lateinit var viewModel: NewestViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -36,23 +34,17 @@ class NewestFragment
 
         newestRecyclerView.also {
             it.adapter = multiTypeAdapter
-            it.layoutManager = LinearLayoutManager(
-                context,
-                LinearLayoutManager.VERTICAL,
-                false
-            )
+            it.layoutManager = LinearLayoutManager(context, LinearLayoutManager.VERTICAL, false)
             it.addItemDecoration(RecyclerViewVerticalDecorator())
         }
 
-        activity?.viewModelStore?.let { viewModelStore ->
-            homeViewModel = ViewModelProvider(
-                viewModelStore,
-                hada().resolve<HomeViewModelFactory>()
-            ).get(HomeViewModel::class.java)
+        viewModel = ViewModelProvider(
+            viewModelStore,
+            hada().resolve<NewestViewModelFactory>()
+        ).get(NewestViewModel::class.java)
 
-            homeViewModel.stateNewest.observe(viewLifecycleOwner) {
-                render(it)
-            }
+        viewModel.stateNewest.observe(viewLifecycleOwner) {
+            render(it)
         }
     }
