+++ app/src/main/res/layout/fragment_upcoming.xml
@@ -33,9 +33,4 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"/>
 
-    <include
-        layout="@layout/custom_view_progress_bar"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
-
 </androidx.constraintlayout.widget.ConstraintLayout>
