+++ data/src/main/java/com/yesferal/hornsapp/data/repository/ConcertRepositoryImpl.kt
@@ -10,8 +10,19 @@ class ConcertRepositoryImpl(
     private val ormDataSource: OrmDataSource,
     private val apiDataSource: ApiDataSource
 ) : ConcertRepository {
+    var concerts: List<Concert>? = null
 
-    override suspend fun getConcerts() = apiDataSource.getConcerts()
+    override suspend fun getConcerts(): Result<List<Concert>> {
+        return concerts?.let {
+            Result.Success(it)
+        } ?: run {
+            val result = apiDataSource.getConcerts()
+            if (result is Result.Success) {
+                concerts = result.value
+            }
+            result
+        }
+    }
 
     override suspend fun getConcert(
         id: String
